feat(cli): add teamsfx validate back
