+++ packages/cli/tests/e2e/bot/ProvisionResourceProjectWithNewBot.tests.ts
@@ -96,7 +96,7 @@ describe("Provision", function () {
     }
 
     // test (validate)
-    await execAsyncWithRetry(`teamsfx manifest validate`, {
+    await execAsyncWithRetry(`teamsfx validate`, {
       cwd: projectPath,
       env: process.env,
       timeout: 0,
