+++ packages/cli/tests/e2e/multienv/TestRemoteHappyPath.tests.ts
@@ -151,7 +151,7 @@ describe("Multi Env Happy Path for Azure", function () {
       }
 
       // validate manifest
-      result = await execAsyncWithRetry(`teamsfx manifest validate --env ${env}`, {
+      result = await execAsyncWithRetry(`teamsfx validate --env ${env}`, {
         cwd: projectPath,
         env: processEnv,
         timeout: 0,
